add dinner utils
